remove remoting, simplify script containers to a singular one
